intelligence/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
brotli==1.2.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
ddgs==9.10.0
distro==1.9.0
fake-useragent==2.2.0
fastapi==0.128.0
filelock==3.20.0
fsspec==2025.12.0
google-auth==2.47.0
google-genai==1.57.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.3.5
packaging==25.0
pillow==12.0.0
primp==0.15.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
safetensors==0.7.0
setuptools==70.2.0
sniffio==1.3.1
socksio==1.0.0
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.22.2
torch==2.9.1+cpu
tqdm==4.67.1
transformers==4.57.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
websockets==15.0.1
